avr: Major serial rewrite. Hope I didn't break it. First part of #2
Serial messages is no longer null terminated and the message's
trailing newline is no longer suppressed. We need to replace (\r)\n
with \0 to keep string functions working until we go all in to Modbus.

Changed data types of buffers to uint8_t* and make the lengths uint8_t
because in practice the buffers will never be larger than 255
bytes. No more is practical anyway with a microcontroller with 2
kilobytes of SRAM. Because of uint8_t pointers, occassional type casts
are necessary to make it work with strcmp and friends.

In case we regret making the length 8-bit, it should be safe to go for
16 bits by changing lengths to uint16_t then.