avr: Make accumulators to drop bits if they run full. Closes #18
This patch defines a certain high water level mark for the accumulator
when the least significant bits of both divident and divisor are
thrown away to give more room. This of course affects the mean
calculation but with so unnoticable amount.

So it doesn't make a great difference in math frees the user from
handling errors related to overflown counters.