avr: Make fill_exception work without knowledge of function code #16
Modbus function codes are not the same as modbus_type_t enum. So we
need to make it independent. The serial buffer is already populated so
it's easy to just stuff it there.

Doing bitwise operation so if it gets called multiple times, the
integer is not overflown but just gives the same result.